% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params}
\alias{get_params}
\title{Get outermost symbol from input parameters inside functions}
\usage{
get_params(x, env)
}
\arguments{
\item{x}{Input argument to be substituted by its symbol}

\item{env}{Top environment to stop searching for \code{x}}
}
\value{
Character vector
}
\description{
Get outermost symbol from input parameters inside functions
}
