% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcfun-C.R
\name{one_iter_intPerm}
\alias{one_iter_intPerm}
\title{test one inter with internal sampling of Fisher-Yates}
\usage{
one_iter_intPerm(res, b, n)
}
\description{
test one inter with internal sampling of Fisher-Yates
}
