% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcfun-C.R
\name{ordervec}
\alias{ordervec}
\title{Order a Vector}
\usage{
ordervec(src, order)
}
\arguments{
\item{src}{a vector of integers to be reordered}

\item{order}{a (zero-indexed) vector of indices}
}
\description{
Generate a new vector from elements in \code{src} ordered by \code{order};
\code{order} may be longer than \code{src} (duplicates allowed).
}
