% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halton.R
\name{randradinv}
\alias{randradinv}
\title{Randomized van der Corput Scrambling Algorithm}
\usage{
randradinv(ind, b)
}
\arguments{
\item{ind}{a vector of integer indices}

\item{b}{a base}
}
\description{
Randomized radical inverse functions for indices in in \code{ind} and for base \code{b}.
The calling routine should set the random seed if reproducibility is desired.
Input is a vector ind of non-negative integer indices and a prime base
\code{b} >= 2. The while loop terminates in floating point arithmetic.
}
