% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halton.R
\name{rhalton}
\alias{rhalton}
\title{Halton sequence}
\usage{
rhalton(n, d, n0 = 0, d0 = 0, singleseed, seedvector)
}
\arguments{
\item{n}{number of points}

\item{d}{number of dimensions}

\item{n0}{initial number point}

\item{d0}{initial number of dimensions}

\item{singleseed}{optional scalar seed (has lower precedence than seedvector)}

\item{seedvector}{optional vector of  d0+d random seeds, one per dimension}
}
\description{
Original R code by Art B. Owen, Stanford University.
Randomly scrambled Halton sequence of n points in d dimensions.
}
\details{
If you already have n0 old points, set n0 to get the next n points.
If you already have d0 old components, set d0 to get the next d inputs.
Get points  n0 + 0:(n-1)  in dimensions  d0 + (1:d)
}
